% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran_pass.R.R
\name{ran_pass}
\alias{ran_pass}
\title{A password generator function}
\usage{
ran_pass(length, type)
}
\arguments{
\item{length}{Length of the password.}

\item{type}{"NL" for passwords with numbers and letters and "N" for passwords with numbers only.}
}
\description{
This function allows you to generate random passwords.
}
\examples{
ran_pass(length = 10, type = "NL")
}
