% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radom_password_function.R.R
\name{random_password}
\alias{random_password}
\title{A Random password generator function
This function allows you to generate radom passwords.}
\usage{
random_password(length, type)
}
\arguments{
\item{length}{Length of the password. With no defaults.
type  "NL" for a password with numbers and letters and "N" for a password with numbers only. With no defaults.}
}
\description{
A Random password generator function
This function allows you to generate radom passwords.
}
\examples{
random_password(length = 10, type = "NL")
}
